Added logic to deal with alternative data format.
NHS England revised 2017-18 releases to include 4 additional columns pertaining to attendances within the 4 hour target, complementary to the columns pertaining to breaches. These new columns appear in the middle of the original column set. Fix is to simply delete them after download, which brings the format back in line with the original.